# NLDD Solver: Skip Unchanged Subdomains Based on Mobility Changes

This PR introduces an a new feature to the NLDD (Nonlinear Domain Decomposition) solver that reduces the time spent on local solves by skipping subdomains that haven't changed significantly between iterations.

Mobility in reservoir simulation represents the ease with which a fluid phase can flow through porous media. It is calculated as the ratio of relative permeability to viscosity for each phase $\alpha$:
```math
\lambda_\alpha = \mathrm{kr}_\alpha / \mu_\alpha
```
Mobility changes indicate significant alterations in fluid flow characteristics, making it an effective indicator for determining when subdomain re-solving is necessary.

##  Details

#### Adaptive Domain Solving
- To determine if we need to solve a domain the relative change in mobility is compared to a previous value and if it is above a given tolerance, the domain is marked for needing solving.
- **Relative change calculation**: 
    `|current_mobility - previous_mobility| / average_cell_mobility > threshold`
- If we do an initial global Newton iteration (default), the change in mobility between before and after this global iteration will be used for the next NLDD iteration.
- If we start with an NLDD iteration all domains will be solved in the first iteration and the change in mobility between before and after this complete NLDD iteration will be used for the next NLDD iteration.
- Domains that required solving in the previous iteration are automatically marked for solving in the next iteration.

####  Reporting
- Added tracking of skipped domains in simulation reports

#### New Parameter
- `--nldd-relative-mobility-change-tol`: Threshold for relative mobility change detection (default: 0.1). By setting this parameter to 0, the mobility thresholding will be disabled.

### Performance Impact
- It has been tested on multiple real reservoir assets, where it reduced the local solve time.
- The reduction in local solve time is most significant for serial runs, for obvious reasons.

### Results on Norne

#### 1 MPI proc

##### Baseline

```
Number of MPI processes:         1
Threads per MPI process:        12
Setup time:                      1.88 s
  Deck input:                    1.25 s
Number of timesteps:           247
Simulation time:               143.50 s
  Assembly time:                12.50 s (Wasted: 0.0 s; 0.0%)
    Well assembly:               1.15 s (Wasted: 0.0 s; 0.0%)
  Linear solve time:            50.67 s (Wasted: 0.0 s; 0.0%)
    Linear setup:               19.21 s (Wasted: 0.0 s; 0.0%)
  Local solve time:             39.29 s (Wasted: 0.0 s; 0.0%)
  Props/update time:             9.04 s (Wasted: 0.0 s; 0.0%)
  Pre/post step:                29.03 s (Wasted: 0.0 s; 0.0%)
  Output write time:             0.38 s
Overall Linearizations:       1257      (Wasted:     0; 0.0%)
Overall Newton Iterations:     590      (Wasted:     0; 0.0%)
Overall Linear Iterations:    1531      (Wasted:     0; 0.0%)`
```
![NORNE_ATW2013_nonlinear_sorted_baseline_NT12](https://github.com/user-attachments/assets/9415faa6-e6a9-4d96-b2a1-3435ba23a944)



##### Mobility thresholding

```
Number of MPI processes:         1
Threads per MPI process:        12
Setup time:                      1.88 s
  Deck input:                    1.26 s
Number of timesteps:           247
Simulation time:               141.01 s
  Assembly time:                12.79 s (Wasted: 0.0 s; 0.0%)
    Well assembly:               1.15 s (Wasted: 0.0 s; 0.0%)
  Linear solve time:            51.64 s (Wasted: 0.0 s; 0.0%)
    Linear setup:               19.70 s (Wasted: 0.0 s; 0.0%)
  Local solve time:             34.92 s (Wasted: 0.0 s; 0.0%)
  Props/update time:             9.14 s (Wasted: 0.0 s; 0.0%)
  Pre/post step:                29.56 s (Wasted: 0.0 s; 0.0%)
  Output write time:             0.38 s
Overall Linearizations:       1273      (Wasted:     0; 0.0%)
Overall Newton Iterations:     605      (Wasted:     0; 0.0%)
Overall Linear Iterations:    1547      (Wasted:     0; 0.0%)
```

![NORNE_ATW2013_nonlinear_sorted_mob_thres_NT12](https://github.com/user-attachments/assets/74d20ce9-57a2-49ba-adbe-dbc1bb92e6a1)


#### 12 MPI proc

##### Baseline

```
Number of MPI processes:        12
Threads per MPI process:         1
Setup time:                      2.63 s
  Deck input:                    1.64 s
Number of timesteps:           247
Simulation time:                91.04 s
  Assembly time:                15.04 s (Wasted: 0.0 s; 0.0%)
    Well assembly:               0.84 s (Wasted: 0.0 s; 0.0%)
  Linear solve time:            27.88 s (Wasted: 0.0 s; 0.0%)
    Linear setup:               12.02 s (Wasted: 0.0 s; 0.0%)
  Local solve time:             17.32 s (Wasted: 0.0 s; 0.0%)
  Props/update time:             7.04 s (Wasted: 0.0 s; 0.0%)
  Pre/post step:                16.99 s (Wasted: 0.0 s; 0.0%)
  Output write time:             4.19 s
Overall Linearizations:       1409      (Wasted:     0; 0.0%)
Overall Newton Iterations:     675      (Wasted:     0; 0.0%)
Overall Linear Iterations:    2154      (Wasted:     0; 0.0%)
```
![NORNE_ATW2013_nonlinear_sorted_baseline_NP12](https://github.com/user-attachments/assets/d11e854f-be47-4f86-806e-e4849bf6105e)


##### Mobility thresholding

```
Number of MPI processes:        12
Threads per MPI process:         1
Setup time:                      2.61 s
  Deck input:                    1.57 s
Number of timesteps:           247
Simulation time:                90.13 s
  Assembly time:                15.43 s (Wasted: 0.0 s; 0.0%)
    Well assembly:               0.83 s (Wasted: 0.0 s; 0.0%)
  Linear solve time:            27.62 s (Wasted: 0.0 s; 0.0%)
    Linear setup:               11.83 s (Wasted: 0.0 s; 0.0%)
  Local solve time:             16.04 s (Wasted: 0.0 s; 0.0%)
  Props/update time:             7.13 s (Wasted: 0.0 s; 0.0%)
  Pre/post step:                17.19 s (Wasted: 0.0 s; 0.0%)
  Output write time:             4.14 s
Overall Linearizations:       1377      (Wasted:     0; 0.0%)
Overall Newton Iterations:     660      (Wasted:     0; 0.0%)
Overall Linear Iterations:    2115      (Wasted:     0; 0.0%)
```
![NORNE_ATW2013_nonlinear_sorted_mob_thres_NP12](https://github.com/user-attachments/assets/d435eaa5-5004-4388-a34b-6854c463c405)