@atgeirr 2025-06-25
> * `nldd_relative_mobility_change_tol_`: Threshold for relative mobility change detection (default: 0.1). By setting this parameter to 0, the mobility thresholding will be disabled.

For the manual: the name of the command line option is `--nldd-relative-mobility-change-tol`