@jakobtorben 2025-06-25
Good clarification Atgeirr, I updated the comment to the actual CLI name to make this clearer.